+++ packages/core/src/exports/internal.ts
@@ -27,9 +27,13 @@ export type { SelectChains } from '../types/chain.js'
 
 export type {
   ChainIdParameter,
+  ConfigParameter,
   ConnectorParameter,
+  EnabledParameter,
   ScopeKeyParameter,
 } from '../types/properties.js'
+
+export type { QueryParameter } from '../types/query.js'
 
 export type {
   Compute,